requirements.txt

apscheduler
mecab-python3
numpy
pandas
requests
requests-oauthlib
tweepy==4.9.0
